requirements.txt
gradio==4.13.0
fal-client==0.0.28
pillow==10.0.1
requests==2.31.0 